support i64 instructions

+++ instr_i32eq.go
@@ -22,7 +22,7 @@ func (instr *InstrI32Eq) Opcode() Opcode {
 func (instr *InstrI32Eq) Perform(ctx context.Context, rt *Runtime) (*Label, error) {
 	return nil, relop(rt, ValTypeI32, func(v1, v2 *Val) (*Val, error) {
 		result := uint32(0)
-		if int32(v1.MustGetI32()) == int32(v2.MustGetI32()) {
+		if v1.MustGetI32() == v2.MustGetI32() {
 			result = uint32(1)
 		}
 		return NewValI32(result), nil
